requirements.txt
aws-cdk-lib==2.179.0
constructs>=10.0.0,<11.0.0
requests==2.32.3
python-dotenv==1.0.1
aws-cdk.aws-scheduler-alpha==2.179.0a0
aws-cdk.aws-scheduler-targets-alpha==2.179.0a0